refactor(notifications): split Windows toast backend
Move the toasted Windows backend into a focused module while keeping toast_notifier as the compatibility and platform-selection surface.

Tested: uv run pytest -q -n 0 --tb=short tests/test_toasted_windows_notifier.py tests/test_windows_app_user_model_id.py tests/test_alert_notification_system.py tests/test_weather_notifier_unified_backend.py tests/test_notification_test.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
